% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build_csv_spec}
\alias{build_csv_spec}
\title{Build a spec from a csv file.}
\usage{
build_csv_spec(file, parsers = NULL, col_names = TRUE,
  quote = "\\"", na_strings = "NA", n = 0, skip = 0, comment_char = "", double_escape = FALSE, backslash_escape = FALSE)
}
\arguments{
\item{n}{If not 0, parse at most this many lines from the file.}

\item{skip}{Skip this many lines before beginning parsing.}

\item{quote}{Quoting character. Quotes are not preserved in the output.}

\item{double_escape}{Does doubling the quoting character escape it?}

\item{backslash_escape}{Do backslashes escape the following character?}

\item{parsers}{A list of parsers used to control how each column is turned
  into an R data frame. If \code{NULL}, \code{\link{guess_parsers}()} will
  do its best to print the correct format.

  Optionally, \code{parsers} can be a named list. If \code{header = TRUE},
  then parsers will be matched to column names, and any unmatched columns
  will not be included in the output. If \code{header = FALSE}, the names
  be used to name the output columns, and there must be one parser per
  column.}

\item{na_strings}{A character vector of providing strings which should be
converted to NA in R.}
}
\description{
Builds a complete specification to parse a csv file, including guessing
the column parsers if they're not suppled.
}
\keyword{internal}

