% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/col_names.R
\name{col_names_standardise}
\alias{col_names_standardise}
\title{Standardise column names}
\usage{
col_names_standardise(col_names, header)
}
\arguments{
\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector
  of column names.

  If \code{TRUE}, the first row of the input will be used as the column
  names, and will not be included in the data frame. If \code{FALSE}, column
  names will be generated automatically: X1, X2, X3 etc.

  If \code{col_names} is a character vector, the values will be used as the
  names of the columns, and the first row of the input will be read into
  the first row of the output data frame.}
}
\description{
Standardise column names
}
\keyword{internal}

