% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collectors.R
\name{collector}
\alias{col_character}
\alias{col_double}
\alias{col_euro_double}
\alias{col_integer}
\alias{col_logical}
\alias{col_skip}
\alias{collector}
\title{Column parsing specifications.}
\usage{
col_character()

col_integer()

col_double()

col_euro_double()

col_logical()

col_skip()
}
\description{
Column parsing specifications.
}

