% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{collector_guess}
\alias{collector_guess}
\title{Guess the type of a vector.}
\usage{
collector_guess(x, locale = default_locale())
}
\arguments{
\item{x}{Character vector of elements to guess.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link{locale}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\description{
Guess the type of a vector.
}
\examples{
collector_guess(c("1", "2", "3"))
collector_guess(c("T", "F", "T"))
collector_guess(c("a", "b", "c"))
}
\keyword{internal}

