% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_lines}
\alias{count_lines}
\title{Count the number of lines in a file.}
\usage{
count_lines(input)
}
\arguments{
\item{input}{Either a path to a file, or a \code{connection}.}
}
\description{
Count the number of lines in a file.
}

