% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-symbols.R
\name{date_names}
\alias{date_names}
\alias{date_names_locale}
\alias{date_names_locales}
\title{Create or retrieve date names}
\usage{
date_names(mon, mon_ab, day, day_ab, am_pm)

date_names_locale(locale, as_ascii = FALSE)

date_names_locales()
}
\arguments{
\item{mon,mon_ab}{Full and abbreviated month names}

\item{day,day_ab}{Full and abbreviated week day names}

\item{am_pm}{Names used for AM and PM.}

\item{locale}{A BCP 47 locale, made up of a languge and a region,
e.g. \code{en_US} for American English. See \code{date_names_locales()}
for a complete list of available locales.}

\item{as_ascii}{Should accents be stripped and the names converted to ASCII?
This is useful if you're dealing with ASCII data where the correct
spellings have been lost. Requires the \pkg{stringi} package.}
}
\description{
When parsing dates, you often need to know how weekdays of the week and
months are represented as text. This pair of functions allows you to either
create your own, or retrieve from a standard list. The standard list is
derived from ICU (\url{http://site.icu-project.org}) via the stringi package.
}
\examples{
date_names_locale("en")
date_names_locale("ko")

date_names_locale("fr")
if (require("stringi")) {
  date_names_locale("fr", as_ascii = TRUE)
}
}

