% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edition.R
\name{edition_get}
\alias{edition_get}
\title{Retrieve the currently active edition}
\usage{
edition_get()
}
\value{
An integer corresponding to the currently active edition.
}
\description{
Retrieve the currently active edition
}
