% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edition.R
\name{edition_set}
\alias{edition_set}
\title{Set the edition used by readr}
\usage{
edition_set(edition)
}
\arguments{
\item{edition}{An integer used to set the readr edition.}
}
\value{
The value of the edition set (invisibly).
}
\description{
Set the edition used by readr
}
