% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encoding.R
\name{guess_encoding}
\alias{guess_encoding}
\title{Guess encoding of file.}
\usage{
guess_encoding(file, n_max = 10000, threshold = 0.2)
}
\arguments{
\item{n_max}{Number of lines to read. If \code{n} is -1, all lines in
file will be read.}

\item{threshold}{Only report guesses above this threshold of certainty.}
}
\description{
Uses \code{\link[stringi]{stri_enc_detect}}: see the documentation there
for caveats.
}

