% Generated by roxygen2 (4.0.1): do not edit by hand
\name{field_spec}
\alias{field_spec}
\title{Field parsing specification for delimited files.}
\usage{
field_spec(delim,
  quote = "\\"", collapse = FALSE, backslash_escape = FALSE, double_escape = FALSE, strict = FALSE)
}
\arguments{
\item{delim}{Delimiter between fields}

\item{quote}{Quoting character. Quotes are not preserved in the output.}

\item{collapse}{Collapse multiple delimiters into one. (Mostly useful when
the delimiter is a space).}

\item{backslash_escape}{Do backslashes escape the following character?}

\item{double_escape}{Does doubling the quoting character escape it?}

\item{strict}{If \code{TRUE}, will throw an error if there are any
unterminated strings or quotes.}
}
\description{
Field parsing specification for delimited files.
}

