% Generated by roxygen2 (4.0.1): do not edit by hand
\name{file_spec}
\alias{delim_spec}
\alias{file_spec}
\alias{fwf_spec}
\title{File parsing specification.}
\usage{
delim_spec(line, field, column)

fwf_spec(widths = NULL, start = NULL, end = NULL, column = NULL)
}
\arguments{
\item{line,field,column}{Line, field and column specification as created by
\code{\link{line_spec}}, \code{\link{field_spec}} and \code{column_spec}.}

\item{widths,start,end}{Either specify the column positions with their
widths, or by providing start and ending positions.}
}
\description{
Parsing a delimited file is naturally composed of three pieces:
parsing a file into lines, parsing each line in to fields, then
converting each field into an R vector, forming the columns of a data frame.
Parsing a fixed width file does not decompose in this way, because while
a delimted file is ragged on disk, a fixed width file is a 2d grid.
}

