% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spec-build.R
\name{guess_column_spec}
\alias{guess_column_spec}
\title{Guess which parsers are needed.}
\usage{
guess_column_spec(fields, parsers = NULL, col_names = TRUE,
  na_strings = "NA", ignore_whitespace = TRUE)
}
\arguments{
\item{fields}{List of character vectors: each vector represents one
lines worth of fields.}
}
\description{
This operation is relatively expensive, so it's only done on a small
sample of the total data frame. This makes it prone to error if your
data formats vary, but since the individual parsers are strict (once
identified) you'll quickly find out about it, and you can specify the
parsers exactly yourself.
}
\keyword{internal}

