% Generated by roxygen2 (4.0.1): do not edit by hand
\name{guess_parsers}
\alias{guess_parsers}
\title{Guess which parsers are needed.}
\usage{
guess_parsers(fields)
}
\arguments{
\item{fields}{List of character vectors: each vector represents one
lines worth of fields.}
}
\description{
Guess which parsers are needed.
}
\keyword{internal}

