% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Input directly into a string.}
\usage{
input(file, text)
}
\arguments{
\item{text}{A character or raw vector. If a character vector, only the
first element is used.}
}
\description{
Input directly into a string.
}
\examples{
input(file.path(R.home(), "COPYING"))
input(text = "abc")
input(text = charToRaw("abc"))
}

