% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spec.R
\name{line_spec}
\alias{line_spec}
\title{Line parsing specification}
\usage{
line_spec(skip = 0, n = 0, comment = "", delim = "\\n")
}
\arguments{
\item{skip}{Skip this many lines before beginning parsing.}

\item{n}{If not 0, parse at most this many lines from the file.}

\item{comment}{If not \code{""}, all content after this character
is ignored.}

\item{delim}{Delimiter between lines.}
}
\description{
Line parsing specification
}

