% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{melt_delim}
\alias{melt_delim}
\alias{melt_csv}
\alias{melt_csv2}
\alias{melt_tsv}
\title{Return melted data for each token}
\usage{
melt_delim(file, delim, quote = "\\"", escape_backslash = FALSE,
  escape_double = TRUE, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = FALSE, skip = 0,
  n_max = Inf)

melt_csv(file, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, quote = "\\"", comment = "", trim_ws = TRUE,
  skip = 0, n_max = Inf)

melt_csv2(file, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, quote = "\\"", comment = "", trim_ws = TRUE,
  skip = 0, n_max = Inf)

melt_tsv(file, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, quote = "\\"", comment = "", trim_ws = TRUE,
  skip = 0, n_max = Inf)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[=clipboard]{clipboard()}} will read from the system clipboard.}

\item{delim}{Single character used to separate fields within a record.}

\item{quote}{Single character used to quote strings.}

\item{escape_backslash}{Does the file use backslashes to escape special
characters? This is more general than \code{escape_double} as backslashes
can be used to escape the delimiter character, the quote character, or
to add special characters like \code{\\n}.}

\item{escape_double}{Does the file escape quotes by doubling them?
i.e. If this option is \code{TRUE}, the value \code{""""} represents
a single quote, \code{\"}.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{quoted_na}{Should missing values inside quotes be treated as missing
values (the default) or strings.}

\item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}

\item{trim_ws}{Should leading and trailing whitespace be trimmed from
each field before parsing it?}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of records to read.}
}
\description{
For certain non-rectangular data format, it can be useful to parse the data
into a melted format where each row is a single token.
}
