% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_fwf.R
\name{melt_fwf}
\alias{melt_fwf}
\title{Return melted data for each token in a fixed width file}
\usage{
melt_fwf(file, col_positions, locale = default_locale(), na = c("", "NA"),
  comment = "", trim_ws = TRUE, skip = 0, n_max = Inf)
}
\description{
For certain non-rectangular data format, it can be useful to parse the data
into a melted format where each row is a single token.
}
