% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_table.R
\name{melt_table}
\alias{melt_table}
\alias{melt_table2}
\title{Return melted data for each token in a whitespace-separated file}
\usage{
melt_table(file, locale = default_locale(), na = "NA", skip = 0,
  n_max = Inf, guess_max = min(n_max, 1000), comment = "")

melt_table2(file, locale = default_locale(), na = "NA", skip = 0,
  n_max = Inf, comment = "")
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[=clipboard]{clipboard()}} will read from the system clipboard.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of records to read.}

\item{guess_max}{Maximum number of records to use for guessing column types.}

\item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}
}
\description{
Return melted data for each token in a whitespace-separated file
}
