% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na_strings.R
\name{na_strings}
\alias{na_strings}
\title{Missing strings}
\usage{
na_strings(data_creator = c("generic", "excel", "sas", "stata", "sql"))
}
\arguments{
\item{data_creator}{A string giving the data creation software}
}
\value{
A character vector.
}
\description{
Missing string by data creation software (Excel, SAS, Stata, SQL currently
supported).  Use this to set the \code{na.strings} argument in
\code{\link{read_table}}.
}
\examples{
na_strings("excel")
}

