% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parser}
\alias{character_parser}
\alias{double_parser}
\alias{factor_parser}
\alias{integer_parser}
\alias{logical_parser}
\alias{parser}
\alias{skip_parser}
\title{Define how to parse each column}
\usage{
character_parser(trim = FALSE)

integer_parser()

double_parser()

logical_parser()

factor_parser(levels, ordered = FALSE)

skip_parser()
}
\arguments{
\item{trim}{If \code{TRUE}, will trim off any leading or trailing whitespace.}

\item{levels}{A character vector listing all valid levels of the factor.}

\item{ordered}{Create an ordered factor?}
}
\description{
Define how to parse each column
}

