% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problems}
\alias{problems}
\alias{stop_for_problems}
\title{Retrieve parsing problems}
\usage{
problems(x = .Last.value)

stop_for_problems(x)
}
\arguments{
\item{x}{A data frame (from \verb{read_*()}) or a vector (from \verb{parse_*()}).}
}
\value{
A data frame with one row for each problem. Columns include:
\item{row,col}{Row and column of problem}
\item{expected}{What readr expected to find}
\item{actual}{What it actually got}

When using the edition 2 parser (the default for \verb{read_*()} functions),
the result also includes a \code{file} column identifying the source file.
}
\description{
Readr functions will only throw an error if parsing fails in an unrecoverable
way. However, there are lots of potential problems that you might want to
know about - these are stored in the \code{problems} attribute of the
output, which you can easily access with this function.
\code{stop_for_problems()} will throw an error if there are any parsing
problems: this is useful for automated scripts where you want to throw
an error as soon as you encounter a problem.
}
\examples{
# parse_*() functions return problems with row, col, expected, actual
x <- parse_integer(c("1X", "blah", "3"))
problems(x)

# When there are no problems, the result is an empty data frame
y <- parse_integer(c("1", "2", "3"))
problems(y)

# read_*() functions also include a file column
z <- read_csv(I("x\n1\n2\nb"), col_types = "d", show_col_types = FALSE)
problems(z)
}
