% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{Read a csv file into a data frame.}
\usage{
read_csv(path, col_types, col_names, na = "NA", text = NULL)
}
\arguments{
\item{path}{Either a path to a file, or a connection. Reading directly
from a file is most efficient.}

\item{col_types}{A list of types used to control how each column is turned
  into an R data frame. The elements of \code{col_types} can be named to match
  the columns. Any columns not included will not be included in the output
  data frame.

  Alternatively, you can use string where each character represents
  one column. c = character, d = double, i = integer, l = logical and
  \code{_} skips the column.}

\item{col_names}{A character vector naming the columns.}

\item{na}{String to use for missing values.}

\item{text}{A character or raw vector. If a character vector, only the
first element is used.}
}
\description{
Read a csv file into a data frame.
}
\examples{
read_csv(text = "1,2\\n3,4", col_types = "ii", col_names = c("x", "y"))
}

