% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table2}
\alias{read_table2}
\title{Read whitespace-separated columns into a tibble}
\usage{
read_table2(
  file,
  col_names = TRUE,
  col_types = NULL,
  locale = default_locale(),
  na = "NA",
  skip = 0,
  n_max = Inf,
  guess_max = min(n_max, 1000),
  progress = show_progress(),
  comment = "",
  skip_empty_rows = TRUE
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because we renamed it to \code{\link[=read_table]{read_table()}} and
removed the old \code{read_table} function, which was too strict for most cases
and was analogous to just using \code{read_fwf()}.
}
\keyword{internal}
