\docType{package}
\name{readR}
\alias{readR}
\alias{readR-package}
\title{Utilities for data import}
\description{
  A collection of utilities for reading and importing data into R by performing (usually small) manipulations of
  data structures such as data frames, matrices and list and automatically determining import parameters.
}
\details{
  \tabular{ll}{
    Package: \tab readR\cr
    Type: \tab Package\cr
    Initial version: \tab 0.1-0\cr
    License: \tab To be used with CORPLOT shiny application only. Please contact the author for more information. \cr
    LazyLoad: \tab yes\cr
  }
  The main function of the package is \code{read.to.list} which reads a number of different file formats into a list of data objects
  such as data frames, depending on the source file.
}
\keyword{package}
