% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readr_threads}
\alias{readr_threads}
\title{Determine how many threads readr should use when processing}
\usage{
readr_threads()
}
\description{
The number of threads returned can be set by
\itemize{
\item The global option \code{readr.num_threads}
\item The environment variable \code{VROOM_THREADS}
\item The value of \code{\link[parallel:detectCores]{parallel::detectCores()}}
}
}
