% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_show_types}
\alias{should_show_types}
\title{Determine whether column types should be shown}
\usage{
should_show_types()
}
\description{
Column types are shown unless
\itemize{
\item They are disabled by setting \code{options(readr.show_col_types = FALSE)}
\item The column types are supplied with the \code{col_types} argument.
}
}
