% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_progress}
\alias{show_progress}
\title{Determine progress bars should be shown}
\usage{
show_progress()
}
\description{
Progress bars are shown \emph{unless} one of the following is \code{TRUE}
\itemize{
\item The bar is explicitly disabled by setting \code{options(readr.show_progress = FALSE)}
\item The code is run in a non-interactive session (\code{interactive()} is \code{FALSE}).
\item The code is run in an RStudio notebook chunk.
\item The code is run by knitr / rmarkdown.
}
}
