% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.R
\name{write_rds}
\alias{write_compressed_rds}
\alias{write_rds}
\title{Write a single R object to file}
\usage{
write_rds(x, path)

write_compressed_rds(x, path, compress_fun = xzfile, compress_amount = 9L,
  ...)
}
\arguments{
\item{x}{R object to write to serialise.}

\item{path}{Path to write to.}

\item{compress_fun}{Compression method to use one of \code{guess},
\code{xzfile}, \code{bzfile}, \code{gzfile}.}

\item{compress_amount}{Level of compression must be in 0:9L}

\item{...}{dots}
}
\description{
Light wrappers around \code{\link{saveRDS}} with explicit naming. \code{write_rds}
never compresses output for speed. \code{write_compressed_rds} tries to guess
the best type of compression for your object to obtain the smallest file size.
}
\section{guess compression}{

N.B. this is not yet implemented but there some empirical testing is taking
place to come up with something useful.
}
\examples{
## write_rds(mtcars, "mtcars.rds")
## write_compressed_rds(mtcars, "mtcars.rds", compress_fun = gzfile)
}

