% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collectors.R
\name{collector}
\alias{col_character}
\alias{col_datetime}
\alias{col_double}
\alias{col_euro_double}
\alias{col_factor}
\alias{col_integer}
\alias{col_logical}
\alias{col_numeric}
\alias{col_skip}
\alias{collector}
\title{Column parsing specifications.}
\usage{
col_character()

col_integer()

col_double()

col_euro_double()

col_numeric()

col_logical()

col_datetime(format = "", tz = "UTC")

col_factor(levels, ordered = FALSE)

col_skip()
}
\arguments{
\item{format}{Format string passed on to the \code{\link{parse_date}}.}

\item{tz}{Default time zone to use for input and output (format).}

\item{levels}{Character vector providing set of allowed levels.}

\item{ordered}{Is it an ordered factor?}
}
\description{
Column parsing specifications.
}

