% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_types.R
\name{cols}
\alias{cols}
\alias{cols_only}
\title{Create column specification}
\usage{
cols(..., .default = col_guess())

cols_only(...)
}
\arguments{
\item{...}{Either column objects created by \code{col_*()}, or their
abbreviated character names. If you're only overriding a few columns,
it's best to refer to columns by name. If not named, the column types
must match the column names exactly.}

\item{.default}{Any named columns not explicitly overridden in \code{...}
will be read with this column type.}
}
\description{
Create column specification
}
\examples{
cols(a = col_integer())
cols_only(a = col_integer())

# You can also use the standard abreviations
cols(a = "i")
cols(a = "i", b = "d", c = "_")

# You can also use multiple sets of column definitions by combining
# them like so:

t1 <- cols(
  column_one = col_integer(),
  column_two = col_number())

t2 <- cols(
 column_three = col_character())

t3 <- t1
t3$cols <- c(t1$cols, t2$cols)
t3
}
