% Generated by roxygen2 (4.0.1): do not edit by hand
\name{column_spec}
\alias{column_spec}
\title{Column parsign specification.}
\usage{
column_spec(parsers, col_names, na_strings = "NA")
}
\arguments{
\item{parsers}{A list of parsers used to control how each column is turned
  into an R data frame. If \code{NULL}, \code{\link{guess_parsers}()} will
  do its best to print the correct format.

  Optionally, \code{parsers} can be a named list. If \code{header = TRUE},
  then parsers will be matched to column names, and any unmatched columns
  will not be included in the output. If \code{header = FALSE}, the names
  be used to name the output columns, and there must be one parser per
  column.}

\item{header}{Does the first line of the file specify column names?
If \code{TRUE}, the first row will be used to define column
names. If \code{FALSE}, columns will be named given names X1-Xn, unless
\code{parsers} is a named list.}

\item{na_strings}{A character vector of providing strings which should be
converted to NA in R.}
}
\description{
Column parsign specification.
}

