% Generated by roxygen2 (4.0.1): do not edit by hand
\name{column_spec}
\alias{column_spec}
\title{Column parsign specification.}
\usage{
column_spec(parsers, col_names, na_strings = "NA")
}
\arguments{
\item{parsers}{A list of parsers used to control how each column is turned
into an R data frame. The elements of \code{parsers} can be named to match
the columns. Any columns not included will not be included in the output
data frame.}

\item{col_names}{A character vector naming the columns.}

\item{na_strings}{A character vector of providing strings which should be
converted to NA in R.}
}
\value{
A list:
\item{parsers}{A list of parsers, in the same order as col_names}
\item{col_names}{}
\item{na_strings}{}
}
\description{
Column parsign specification.
}

