% Generated by roxygen2 (4.0.1): do not edit by hand
\name{count_lines}
\alias{count_lines}
\title{Count the number of lines in a file.}
\usage{
count_lines(input)
}
\arguments{
\item{input}{Either a path to a file, or a \code{connection}.}

\item{header}{If \code{TRUE}, subtract one line to account for the
header.}
}
\description{
Count the number of lines in a file.
}

