% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{datasource}
\alias{datasource}
\title{Create a source object.}
\usage{
datasource(file, skip = 0, comment = "", encoding = NULL)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path).}

\item{skip}{Number of lines to skip before reading data.}

\item{comment}{A string or character vector used to identify comments. Any
text after the comment characters will be silently ignored.}

\item{encoding}{The text encoding of the data given in \code{file}. If \code{NULL} is
passed (the default), then "UTF-8" is used for files,
"bytes" is used for connections and raw vectors, and the
string Encoding will be used for literal data strings.}
}
\description{
Create a source object.
}
\examples{
# Literal csv
datasource("a,b,c\\n1,2,3")
datasource(charToRaw("a,b,c\\n1,2,3"))

# Strings
datasource(readr_example("mtcars.csv"))
datasource(readr_example("mtcars.csv.bz2"))
datasource(readr_example("mtcars.csv.zip"))
\dontrun{
datasource("https://github.com/tidyverse/readr/raw/master/inst/extdata/mtcars.csv")
}

# Connection
con <- rawConnection(charToRaw("abc\\n123"))
datasource(con)
close(con)
}
\keyword{internal}
