% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locale.R
\name{locale}
\alias{default_locale}
\alias{locale}
\title{Create locales}
\usage{
locale(date_names = "en", date_format = "\%Y\%.\%m\%.\%d",
  time_format = "\%H:\%M", decimal_mark = ".", grouping_mark = NULL,
  tz = "UTC", encoding = "UTF-8")

default_locale()
}
\arguments{
\item{date_names}{Character representations of day and month names. Either
a string, which is passed on to \code{\link{date_names_locale}}, or
an object created by \code{\link{date_names}}.}

\item{date_format,time_format}{Default date and time formats.}

\item{decimal_mark,grouping_mark}{Symbols used to indicate the decimal
place, and to chunk larger numbers. Decimal mark can only be \code{,} or
\code{.}.}

\item{tz}{Default time zone.}

\item{encoding}{Default encoding. This only affects how the file is
read - readr always converts the output to UTF-8.}
}
\description{
A locale object tries to capture all the defaults that can vary between
countries. You set the locale in once, and the details are automatically
passed on down to the columns parsers. The defaults have been chosen to
match R (i.e. US English) as closely as possible.
}
\examples{
locale()
locale("fr")

# South American locale
locale("es_MX", decimal_mark = ",")
}

