% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{parse_factor}
\alias{col_factor}
\alias{parse_factor}
\title{Parse a character vector into a factor}
\usage{
parse_factor(x, levels, ordered = FALSE, na = c("", "NA"),
  locale = default_locale())

col_factor(levels, ordered = FALSE)
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{levels}{Character vector providing set of allowed levels.}

\item{ordered}{Is it an ordered factor?}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link{locale}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\description{
Parse a character vector into a factor
}
\examples{
parse_factor(c("a", "b"), letters)
}
\seealso{
Other parser: \code{\link{parse_atomic}},
  \code{\link{parse_datetime}}, \code{\link{parse_guess}},
  \code{\link{parse_number}}
}

