% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{parse_factor}
\alias{parse_factor}
\alias{col_factor}
\title{Parse factors}
\usage{
parse_factor(x, levels, ordered = FALSE, na = c("", "NA"),
  locale = default_locale())

col_factor(levels, ordered = FALSE)
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{levels}{Character vector providing set of allowed levels.}

\item{ordered}{Is it an ordered factor?}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link{locale}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\description{
\code{parse_factor} is similar to \code{\link{factor}}, but will generate
warnings if elements of \code{x} are not found in \code{levels}.
}
\examples{
parse_factor(c("a", "b"), letters)

x <- c("cat", "dog", "caw")
levels <- c("cat", "dog", "cow")

# Base R factor() silently converts unknown levels to NA
x1 <- factor(x, levels)

# parse_factor generates a warning & problems
x2 <- parse_factor(x, levels)
}
\seealso{
Other parsers: \code{\link{col_skip}},
  \code{\link{parse_datetime}}, \code{\link{parse_guess}},
  \code{\link{parse_logical}}, \code{\link{parse_number}}
}
