% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{parse_number}
\alias{parse_number}
\alias{col_number}
\title{Parse numbers, flexibly}
\usage{
parse_number(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE)

col_number()
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, grouping mark, and day/month
names.}

\item{trim_ws}{Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from
each field before parsing it?}
}
\value{
A numeric vector (double) of parsed numbers.
}
\description{
This parses the first number it finds, dropping any non-numeric characters
before the first number and all characters after the first number. The
grouping mark specified by the locale is ignored inside the number.
}
\examples{
## These all return 1000
parse_number("$1,000") ## leading `$` and grouping character `,` ignored
parse_number("euro1,000") ## leading non-numeric euro ignored
parse_number("t1000t1000") ## only parses first number found

parse_number("1,234.56")
## explicit locale specifying European grouping and decimal marks
parse_number("1.234,56", locale = locale(decimal_mark = ",", grouping_mark = "."))
## SI/ISO 31-0 standard spaces for number grouping
parse_number("1 234.56", locale = locale(decimal_mark = ".", grouping_mark = " "))

## Specifying strings for NAs
parse_number(c("1", "2", "3", "NA"))
parse_number(c("1", "2", "3", "NA", "Nothing"), na = c("NA", "Nothing"))
}
\seealso{
Other parsers:
\code{\link[=col_skip]{col_skip()}},
\code{\link[=cols]{cols()}},
\code{\link[=cols_condense]{cols_condense()}},
\code{\link[=parse_datetime]{parse_datetime()}},
\code{\link[=parse_factor]{parse_factor()}},
\code{\link[=parse_guess]{parse_guess()}},
\code{\link[=parse_logical]{parse_logical()}},
\code{\link[=parse_vector]{parse_vector()}}
}
\concept{parsers}
