% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{parse_number}
\alias{col_number}
\alias{parse_number}
\title{Extract numbers out of an atomic vector}
\usage{
parse_number(x, na = c("", "NA"), locale = default_locale())

col_number()
}
\arguments{
\item{x}{Character vector of values to parse.}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link{locale}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\description{
This drops any non-numeric characters before or after the first number.
The grouping mark specified by the locale is ignored inside the number.
}
\examples{
parse_number("$1000")
parse_number("1,234,567.78")
}
\seealso{
Other parser: \code{\link{parse_atomic}},
  \code{\link{parse_datetime}}, \code{\link{parse_factor}},
  \code{\link{parse_guess}}
}

