% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_vector}
\alias{parse_vector}
\title{Parse into a vector.}
\usage{
parse_vector(text, tokenizer, collector)
}
\arguments{
\item{text}{String to parse.}
}
\description{
This tokenizes then collects the input values into a single vector.
It's mostly useful for testing
}
\examples{
parse_vector("1,2,3,NA", tokenizer_csv(), col_integer())
parse_vector("1,2,3,NA", tokenizer_csv(), col_double())
parse_vector("1,2,3,NA", tokenizer_csv(), col_character())
parse_vector("1,2,3,NA", tokenizer_csv(), col_skip())

parse_vector("1,2,3,.", tokenizer_csv(na = "."), col_double())

# Invalid values are replaced with missing values with a warning.
parse_vector("1,2,-\\n", tokenizer_csv(), col_double())
parse_vector("1,2,-\\n", tokenizer_csv(), col_integer())
}
\keyword{internal}

