% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_vector}
\alias{parse_vector}
\title{Parse a character vector.}
\usage{
parse_vector(x, collectorSpec)
}
\arguments{
\item{x}{Character vector of elements to parse.}

\item{collector}{Column specification.}
}
\description{
Parse a character vector.
}
\examples{
x <- c("1", "2", "3", NA)
parse_vector(x, col_integer())
parse_vector(x, col_double())
parse_vector(x, col_character())
parse_vector(x, col_skip())

# Invalid values are replaced with missing values with a warning.
x <- c("1", "2", "3", "-")
parse_vector(x, col_double())
}
\keyword{internal}

