% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_catalog.R
\name{read_catalog}
\alias{read_catalog}
\title{Read multiple files from an input catalog}
\usage{
read_catalog(inca, bind = FALSE)
}
\arguments{
\item{inca}{(df) A input catalog (inca).}

\item{bind}{(lgl). Should \code{dplyr::bind_rows()} be called on output?
Defaults to FALSE.}
}
\description{
An input catalog (inca for short?) is a table that catalogs the inputs of a
project. It is a `data.frame` object where each observation/record/row
represents an input: something to be inputed into R relevant to your project.
}
\details{
Defining variables are: \code{in_name} a character string of the human-friendly
name of the input, \code{in_fun} a character string of the name of the
function to use to read the input, and \code{in_args} a named list of the
arguments to be used when reading the input.
}
\examples{
# see the vignette.
}

