% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_catalog.R
\name{read_catalog}
\alias{read_catalog}
\title{Read multiple files from a file catalog}
\usage{
read_catalog(catalog, bind = FALSE)
}
\arguments{
\item{catalog}{df. A file catalog.}

\item{bind}{boolean. Should \code{dplyr::bind_rows()} be called on output.
Defaults to FALSE.}
}
\description{
A file catalog is a \code{data.frame} object (df) where each observation is a
specification of how to read in a particular file. The minimum variables are
\code{file}, \code{read_fun} and \code{dots}.
}
\examples{
# first we make a catalog
 files <- c(system.file("extdata/mtcars.csv", package = "readr"),
           system.file("extdata/mtcars.csv.zip", package = "readr"),
           system.file("extdata/mtcars.csv.bz2", package = "readr"))
 catalog <- dplyr::data_frame(file = files,
                             read_fun = "read_csv",
                             dots = list(list(skip = 1),
                                         list(n_max = 1),
                                        list(col_names = F)))
 read_catalog(catalog)
 read_catalog(catalog, bind = TRUE)
}

