% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_csv}
\alias{read_csv}
\title{Read a csv file}
\usage{
read_csv(file, parsers = NULL, col_names = TRUE, na_strings = "NA", ...)
}
\arguments{
\item{file}{Path to a file, or a connection object.}

\item{...}{Other parameters passed on to \code{\link{build_csv_spec}}.
These are not normally necessary, but you may need them if you have
an unusually formatted csv file.}

\item{parsers}{A list of parsers used to control how each column is turned
into an R data frame. The elements of \code{parsers} can be named to match
the columns. Any columns not included will not be included in the output
data frame.}

\item{col_names}{A character vector naming the columns.}

\item{na_strings}{A character vector of providing strings which should be
converted to NA in R.}
}
\value{
A data frame with additionanl class \code{"tbl_df"}.
}
\description{
Compared to \code{\link{read.csv}} this function is considerably faster.
}

