% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_delim}
\alias{read_csv}
\alias{read_delim}
\title{Read a delimited file.}
\usage{
read_delim(file, delim = NULL, parsers = NULL, col_names = TRUE, ...,
  na_strings = "NA")

read_csv(file, parsers = NULL, col_names = TRUE, ..., na_strings = "NA")
}
\arguments{
\item{file}{Path to a file, or a connection object.}

\item{delim}{Field delimiter.}

\item{...}{Other parameters passed on to \code{\link{build_delim_spec}}.
These are not normally necessary, but maybe be useful if you have
an unusually formatted file.}

\item{parsers}{A list of parsers used to control how each column is turned
  into an R data frame. The elements of \code{parsers} can be named to match
  the columns. Any columns not included will not be included in the output
  data frame.

  Alternatively, you can use string where each character represents
  one column. c = character, d = double, i = integer, l = logical and
  \code{_} skips the column.}

\item{col_names}{A character vector naming the columns.}

\item{na_strings}{A character vector of providing strings which should be
converted to NA in R.}
}
\value{
A data frame with additional class \code{"tbl_df"}.
}
\description{
\code{read_delim} is usually 5-15x faster than \code{\link{read.delim}}, and
provides substantially more feedback about what's going on. It produces a
data frame inline with modern R usage: col names are left as is, factors
are never created by default, and row names are never used. \code{read_csv}
is a version of \code{read_delim} where \code{delim} is set to ",".
}

