% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chunked.R
\name{read_delim_chunked}
\alias{read_csv2_chunked}
\alias{read_csv_chunked}
\alias{read_delim_chunked}
\alias{read_delimited_chunked}
\alias{read_tsv_chunked}
\title{Read a delimited file by chunks}
\usage{
read_delimited_chunked(file, callback, chunk_size = 10000, tokenizer,
  col_names = TRUE, col_types = NULL, locale = default_locale(),
  skip = 0, comment = "", guess_max = min(1000, n_max),
  progress = interactive())

read_csv_chunked(file, callback, chunk_size = 10000, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE, skip = 0,
  guess_max = min(1000, n_max), progress = interactive())

read_csv2_chunked(file, callback, chunk_size = 10000, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE, skip = 0,
  guess_max = min(1000, n_max), progress = interactive())

read_tsv_chunked(file, callback, chunk_size = 10000, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE, skip = 0,
  guess_max = min(1000, n_max), progress = interactive())
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
   (either a single string or a raw vector).

   Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
   be automatically uncompressed. Files starting with \code{http://},
   \code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
   downloaded. Remote gz files can also be automatically downloaded &
   decompressed.

   Literal data is most useful for examples and tests. It must contain at
   least one new line to be recognised as data (instead of a path).}

\item{callback}{A callback function to call on each chunk}

\item{chunk_size}{The number of rows to include in each chunk}

\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector
  of column names.

  If \code{TRUE}, the first row of the input will be used as the column
  names, and will not be included in the data frame. If \code{FALSE}, column
  names will be generated automatically: X1, X2, X3 etc.

  If \code{col_names} is a character vector, the values will be used as the
  names of the columns, and the first row of the input will be read into
  the first row of the output data frame.

  Missing (\code{NA}) column names will generate a warning, and be filled
  in with dummy names \code{X1}, \code{X2} etc. Duplicate column names
  will generate a warning and be made unique with a numeric prefix.}

\item{col_types}{One of \code{NULL}, a \code{\link{cols}} specification, or
  a string. See \code{vignette("column-types")} for more details.

  If \code{NULL}, all column types will be imputed from the first 1000 rows
  on the input. This is convenient (and fast), but not robust. If the
  imputation fails, you'll need to supply the correct types yourself.

  If a column specification created by \code{\link{cols}}, it must contain
  one column specification for each column. If you only want to read a
  subset of the columns, use \code{\link{cols_only}}.

  Alternatively, you can use a compact string representation where each
  character represents one column:
  c = character, i = integer, n = number, d = double,
  l = logical, D = date, T = date time, t = time, ? = guess, or
  \code{_}/\code{-} to skip the column.}

\item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link{locale}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{skip}{Number of lines to skip before reading data.}

\item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}

\item{guess_max}{Maximum number of records to use for guessing column types.}

\item{progress}{Display a progress bar? By default it will only display
in an interactive session. The display is updated every 50,000 values
and will only display if estimated reading time is 5 seconds or more.}

\item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{quoted_na}{Should missing values inside quotes be treated as missing
values (the default) or strings.}

\item{trim_ws}{Should leading and trailing whitespace be trimmed from
each field before parsing it?}
}
\description{
Read a delimited file by chunks
}

