% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{read_file}
\alias{read_file}
\title{Read a file into a string.}
\usage{
read_file(path, text)
}
\arguments{
\item{path}{Either a path to a file, or a connection. Reading directly
from a file is most efficient.}

\item{text}{A character or raw vector. If a character vector, only the
first element is used.}
}
\description{
Read a file into a string.
}
\examples{
read_file(file.path(R.home(), "COPYING"))
read_file(text = "abc")
read_file(text = charToRaw("abc"))
}

