% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_lines}
\alias{read_lines}
\title{Read lines}
\usage{
read_lines(input, n = 0L)
}
\arguments{
\item{input}{Either a path to a file, or a \code{connection}.}

\item{n}{Number of lines to read. If \code{0}, all lines in the file
will be read.}
}
\description{
Read lines
}

