% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_lines.R
\name{read_lines}
\alias{read_lines}
\title{Read lines from a file or string.}
\usage{
read_lines(path, text, n = -1L)
}
\arguments{
\item{path}{Either a path to a file, or a connection. Reading directly
from a file is most efficient.}

\item{text}{A character or raw vector. If a character vector, only the
first element is used.}

\item{n}{Number of lines to read. If \code{n} is -1, all lines in
file will be read. For large files, this may}
}
\value{
A character vector with one element for each line.
}
\description{
Read lines from a file or string.
}
\examples{
read_lines(text = "1\\n\\n2")
tokenize(text = "1\\n\\n2", tokenizer = line_tokenizer())
}

