% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds.R
\name{read_rds}
\alias{read_rds}
\title{Read object from RDS file.}
\usage{
read_rds(path)
}
\arguments{
\item{path}{Path of file}
}
\description{
This is a minimal wrapper around \code{\link{readRDS}} that uses the
same naming scheme as all other functions in \pkg{readr}.
}
\examples{
temp <- tempfile()
write_rds(mtcars, temp)
read_rds(temp)
}

