% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr-package.R
\docType{package}
\name{readr-package}
\alias{readr}
\alias{readr-package}
\title{readr: Read Rectangular Text Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'readr' is to provide a fast and friendly way to read rectangular data (like 'csv', 'tsv', and 'fwf'). It is designed to flexibly parse many types of data found in the wild, while still cleanly failing when data unexpectedly changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://readr.tidyverse.org}
  \item \url{https://github.com/tidyverse/readr}
  \item Report bugs at \url{https://github.com/tidyverse/readr/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
  \item Jim Hester
}

Other contributors:
\itemize{
  \item Romain Francois [contributor]
  \item Shelby Bearrows [contributor]
  \item Posit Software, PBC [copyright holder, funder]
  \item https://github.com/mandreyel/ (mio library) [copyright holder]
  \item Jukka Jylänki (grisu3 implementation) [contributor, copyright holder]
  \item Mikkel Jørgensen (grisu3 implementation) [contributor, copyright holder]
}

}
\keyword{internal}
