% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr.R
\docType{package}
\name{readr-package}
\alias{readr}
\alias{readr-package}
\title{readr: Read Rectangular Text Data}
\description{
The goal of 'readr' is to provide a fast and friendly way to read
rectangular data (like 'csv', 'tsv', and 'fwf'). It is designed to flexibly
parse many types of data found in the wild, while still cleanly failing when
data unexpectedly changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://readr.tidyverse.org}
  \item \url{https://github.com/tidyverse/readr}
  \item Report bugs at \url{https://github.com/tidyverse/readr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Jim Hester \email{jim.hester@rstudio.com}
  \item Romain Francois
}

Other contributors:
\itemize{
  \item R Core Team (Date time code adapted from R) [contributor]
  \item RStudio [copyright holder, funder]
  \item Jukka Jylänki (grisu3 implementation) [contributor, copyright holder]
  \item Mikkel Jørgensen (grisu3 implementation) [contributor, copyright holder]
}

}
\keyword{internal}

