% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{readr_example}
\alias{readr_example}
\title{Get path to readr example}
\usage{
readr_example(path)
}
\arguments{
\item{path}{Name of file}
}
\description{
readr comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
readr_example("challenge.csv")
}
\keyword{internal}
