% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_read_lazy}
\alias{should_read_lazy}
\title{Determine whether reading should be done lazily}
\usage{
should_read_lazy()
}
\description{
Reading in readr is not done lazily unless
\itemize{
\item \code{options(readr.read_lazy = TRUE)}
}
}
