% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_show_types}
\alias{should_show_types}
\title{Determine whether column types should be shown}
\usage{
should_show_types()
}
\description{
Wrapper around \code{getOption("readr.show_col_types")} that implements some fall
back logic if the option is unset. This returns:
\itemize{
\item \code{TRUE} if the option is set to \code{TRUE}
\item \code{FALSE} if the option is set to \code{FALSE}
\item \code{FALSE} if the option is unset and we appear to be running tests
\item \code{NULL} otherwise, in which case the caller determines whether to show
column types based on context, e.g. whether \code{show_col_types} or actual
\code{col_types} were explicitly specified
}
}
