% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_types.R
\name{spec}
\alias{spec}
\title{Extract the column specification from an data frame}
\usage{
spec(x)
}
\arguments{
\item{x}{The data frame object to extract from}
}
\value{
The col_spec object used for that data frame.
}
\description{
Extract the column specification from an data frame
}

