% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_types.R
\name{condense_spec}
\alias{condense_spec}
\alias{spec}
\title{Examine the column specifications for a data frame}
\usage{
condense_spec(x)

spec(x)
}
\arguments{
\item{x}{The data frame object to extract from}
}
\value{
A col_spec object.
}
\description{
\code{\link{spec}} extracts the full column specifications.
\code{\link{condense_spec}} takes a spec object and condenses its definition
by setting the default column type to the most frequent type and only
listing columns with a different type.
}
\examples{
df <- read_csv(system.file("mtcars.csv", package = "readr"))
s <- spec(df)
s

condense_spec(s)
}

