% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{Tokenizers}
\alias{Tokenizers}
\alias{tokenizer_csv}
\alias{tokenizer_line}
\title{Tokenizers.}
\usage{
tokenizer_csv(na = "NA")

tokenizer_line()
}
\arguments{
\item{na}{String to use for missing values.}
}
\description{
Explicitly create tokenizer objects. Usually you will not call these
function, but will instead use one of the use friendly wrappers like
\code{read_csv}.
}
\examples{
tokenizer_csv()
}
\keyword{internal}

