% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{Tokenizers}
\alias{Tokenizers}
\alias{csv_tokenizer}
\title{Tokenizers.}
\usage{
csv_tokenizer(na = "NA")
}
\arguments{
\item{na}{String to use for missing values.}
}
\description{
Explicitly create tokenizer objects. Usually you will not call these
function, but will instead use one of the use friendly wrappers like
\code{read_csv}.
}
\examples{
csv_tokenizer()
}
\keyword{internal}

