% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lines.R
\name{write_lines}
\alias{write_lines}
\title{Write line to a file or string.}
\usage{
write_lines(x, path, na = "NA", append = FALSE)
}
\arguments{
\item{x}{A data frame to write to disk}

\item{path}{Path to write to.}

\item{na}{String used for missing values. Defaults to NA. Missing values
will never be quoted; strings with the same value as \code{na} will
always be quoted.}

\item{append}{If \code{FALSE}, will overwrite existing file. If \code{TRUE},
will append to existing file. In both cases, if file does not exist a new
file is created.}
}
\value{
returns the input \code{x} invisibly.
}
\description{
Write line to a file or string.
}
\examples{
tmp <- tempfile()
write_lines(rownames(mtcars), tmp)

write_lines(airquality$Ozone, tmp, na = "-1")
}

