% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.R
\name{write_rds}
\alias{write_rds}
\title{Write a single R object to file}
\usage{
write_rds(x, path, compress = c("none", "gz", "bz2", "xz"), ...)
}
\arguments{
\item{x}{R object to write to serialise.}

\item{path}{Path to write to.}

\item{compress}{Compression method to use one of \code{c("none", "gz" ,"bz",
"xz")}.}

\item{...}{Additional arguments to connection function. For example, control
the space-time trade-off of different compression methods with
\code{compression}. See \code{\link{connections}} for more details.}
}
\description{
Consistent wrapper around \code{\link{saveRDS}}. \code{write_rds} never
compresses by default as space is generally cheaper than time.
}
\examples{
## write_rds(mtcars, "mtcars.rds")
## write_rds(mtcars, "compressed_mtc.rds", "xz", compression = 9L)
}

